adds files for modularization, adds some events, and coordinates between files
