+++ chatroom.js
@@ -11,6 +11,10 @@ const server = net.createServer();
 const socketPool = {};
 const commands = {};
 
+const logger = require('./modules/logger.js');
+const app = require('./modules/app.js');
+const events = require('./modules/events.js');
+
 server.on('connection', (socket) => {
   let id = uuid();
   socketPool[id] = {
@@ -21,16 +25,26 @@ server.on('connection', (socket) => {
   socket.on('data', (buffer) => dispatchAction(id, buffer));
 });
 
-let parse = (buffer) => {
-  let text = buffer.toString().trim();
-  if ( !text.startsWith('@') ) { return null; }
-  let [command,payload] = text.split(/\s+(.*)/);
-  let [target,message] = payload.split(/\s+(.*)/);
-  return {command,payload,target,message};
-};
+// let parse = (buffer) => {
+//   let text = buffer.toString().trim();
+//   if ( !text.startsWith('@') ) { return null; }
+//   let [command,payload] = text.split(/\s+(.*)/);
+//   let [target,message] = payload.split(/\s+(.*)/);
+//   return {command,payload,target,message};
+// };
+
+events.on('accept-buffer', acceptBuffer);
+
+let entry;
+
+function acceptBuffer ({command,payload,target,message}){
+  entry = {command,payload,target,message};
+  return entry;
+}
 
 let dispatchAction = (userId, buffer) => {
-  let entry = parse(buffer);
+  app.parse(buffer);
+  console.log('entry', entry);
   if ( entry && typeof commands[entry.command] === 'function' ) {
     commands[entry.command](entry, userId);
   }
