+++ modules/events.js
@@ -1,0 +1,4 @@
+const EE = require('events');
+const events = new EE();
+
+module.exports = events;
