+++ package.json
@@ -12,7 +12,8 @@
   "dependencies": {
     "eslint": "^5.11.1",
     "jest": "^23.6.0",
+    "nc": "^1.0.2",
     "node": "^11.5.0",
-    "uuid": "^3.2.1"
+    "uuid": "^3.3.2"
   }
 }
